# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Middletown Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PaymentVerifyInput {
  /** Paystack payment reference */
  reference: string;
  /** Firestore order document ID */
  orderId: string;
}

export interface PaymentVerifyResult {
  success: boolean;
  message: string;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  paidAt?: string | null;
}

export interface ErrorResponse {
  error: string;
}