# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Middletown Shop API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: paystack
    description: Paystack payment operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [paystack]
      summary: Verify a Paystack payment server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    PaymentVerifyInput:
      type: object
      required:
        - reference
        - orderId
      properties:
        reference:
          type: string
          description: Paystack payment reference
        orderId:
          type: string
          description: Firestore order document ID

    PaymentVerifyResult:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
        transactionId:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string