# lib/api-zod/src/generated/types/paymentVerifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Middletown Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerifyInput {
  /** Paystack payment reference */
  reference: string;
  /** Firestore order document ID */
  orderId: string;
}