# lib/api-zod/src/generated/types/paymentVerifyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Middletown Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerifyResult {
  success: boolean;
  message: string;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  paidAt?: string | null;
}